add alls

+++ src/main/resources/db/migration/V2__student.sql
@@ -1,10 +1,10 @@
 CREATE TABLE `studenthyou` (
-	`id` SERIAL NOT NULL COMMENT 'ID',
-	`no` VARCHAR(10),
-	`name` VARCHAR(10),
-	`ent_year` INTEGER(10),
-	`class_num` CHAR(3),
-	`is_attend` BOOLEAN,
-	`school_cd` CHAR(3),
-	PRIMARY KEY (`id`)
+    `id` SERIAL NOT NULL COMMENT 'ID',
+    `no` VARCHAR(10),
+    `name` VARCHAR(10),
+    `ent_year` INTEGER,
+    `class_num` CHAR(3),
+    `is_attend` BOOLEAN,
+    `school_cd` CHAR(3),
+    PRIMARY KEY (`id`)
 ) ENGINE=InnoDB;
